+++ src/stdlib/conv/strtol.c
@@ -3,14 +3,14 @@
 #include <errno.h>
 
 long strtol(const char * restrict nptr, char ** restrict endptr, int base) {
-	uint64_t ret = 0;
+	unsigned long long ret = 0;
 	int sign = scan_int(nptr, endptr, base, &ret);
 	if (sign < 0) {
 		errno = -sign;
 		return sign == -ERANGE ? (ret ? LONG_MIN : LONG_MAX) : ret;
 	}
 	if (sign) {
-		if (ret > -(uint64_t)LONG_MIN) {
+		if (ret > -(unsigned long long)LONG_MIN) {
 			errno = ERANGE;
 			return LONG_MIN;
 		}
