+++ src/stdlib/conv/strtoll.c
@@ -3,14 +3,14 @@
 #include <errno.h>
 
 long long strtoll(const char * restrict nptr, char ** restrict endptr, int base) {
-	uint64_t ret = 0;
+	unsigned long long ret = 0;
 	int sign = scan_int(nptr, endptr, base, &ret);
 	if (sign < 0) {
 		errno = -sign;
 		return sign == -ERANGE ? (ret ? LLONG_MIN : LLONG_MAX) : ret;
 	}
 	if (sign) {
-		if (ret > -(uint64_t)LLONG_MIN) {
+		if (ret > -(unsigned long long)LLONG_MIN) {
 			errno = ERANGE;
 			return LLONG_MIN;
 		}
